Remove IncludeCompositeRoot functionality and add unreferenced code attributes
The IncludeCompositeRoot functionality used to control the serialization of the root activity of composite workflows has been removed. Also, [RequiresUnreferencedCode] attributes have been added to serialization-related methods hinting that the type of the model must be known at compile time. Commit ensures that serialization changes doesn't break the existing application, and the attribute helps in optimizing the codebase's performance and security.